Implement options menu in all Activities
MenuActivity contains an options menu that can bring up the
AboutActivity, or a QuoteActivity with a random quote, or the user's
most recently viewed Quote.

All other Activities now extend MenuActivity. This allows them all to
use the same menu, while keeping the menu code in a single file, for
ease of modification or use.

TODO: implement code for "Last" button.