Persist QuoteActivity category title correctly.
Previously, SharedPreferences saved all text in quoteTitleTV on activity
pause. This meant that it would append an additional "Quote about " at
the beginning of the title whenever the quote was loaded through the
"Last quote viewed" option in the menu.

Current implementation pulls the category title directly from the extras
in QuoteActivity's intent. This allows the title to be displayed
properly on load, and prevents it from growing unexpedtedly in size if a
quote is repeatedly closed and reopened via options.