Prevent "Last" option fetching null database data.
Previous implementation of QuoteActivity caused accidental increase of
quoteID every time a quote was closed and re-opened. This caused the
database to request the quote with an ID 1 higher than intended every
time, eventually causing it to request an ID that did not exist.

quoteID originally ranged from 0 to 3 due to it being pulled from a
zero-indexed list in the adapter. It now has +1 appended to it within
CustomAdapter when being saved to intent extras, instead of handling the
conversion externally.